requirements.txt
django>5.1,<5.2
environs[django] == 11.0.0
pillow
djangorestframework
markdown
django-filter